Add SystemConfigurationManager that supports user provided Configuration
